Fix panic in useragent.WithUserAgentExtra
Fixes #135

The upstream `useragent` library panics when receiving a semver beginning with
"v". This call to `useragent` was added in #134 to address requests that we
pass a Pulumi user agent.

This panic was missed by testing due to pulumi/pulumi#12981, panics in a
provider during ProgramTest result in an erroneous passing result.

Tracing the source of the panic, we see `useragent.WithUserAgentExtra` calls data.With here.

https://github.com/databricks/databricks-sdk-go/blob/v0.9.0/useragent/user_agent.go#L42-L46
```go
// WithUserAgentExtra sets per-process extra user agent data,
// which integration developers have agreed with Databricks.
func WithUserAgentExtra(key, value string) {
	extra = extra.With(key, value)
}
```

The `With` method then performs two checks, and if either fails it panics:

https://github.com/databricks/databricks-sdk-go/blob/v0.9.0/useragent/user_agent.go#L88-L96
```
// With always uses the latest value for a given alphanumeric key.
// Panics if key or value don't satisfy alphanumeric or semver format.
func (d data) With(key, value string) data {
  if err := matchAlphanum(key); err != nil {
    panic(err)
  }
  if err := matchAlphanumOrSemVer(value); err != nil {
    panic(err)
  }
  // ...
}
```